pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agency-toolbox"
version = "0.1.0"
description = "Internal tool for environmental consulting data management"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your.email@agency.com" }
]
dependencies = [
    "aiosqlite>=0.20.0", # For asynchronous SQLite support
    "bcrypt",
    "fastapi>=0.110.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.27.0",
]

[project.scripts]
dev = "uvicorn.main:main"
setup-db = "app.scripts.db:initialize"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pyright]
# Helps VS Code's type checking ignore your L: drive source
venvPath = "."
venv = ".venv"